requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pytest-django==4.6.0
python-decouple==3.8
psycopg2-binary==2.9.9
drf-spectacular==0.26.5
flake8-django==1.4
